+++ src/builders.ts
@@ -1,6 +1,8 @@
 import { types as t } from '@babel/core'
 import { Expression } from '@babel/types'
 
+import { ResponsiveStyles } from './utils'
+
 /**
  * Builds an object property with the provided parameters.
  *
@@ -13,3 +15,18 @@ export const buildObjectProperty = (
   identifier: string,
   expression: Expression
 ) => t.objectProperty(t.identifier(identifier), expression)
+
+/**
+ * Provided keyed and grouped repsonsive styles, creates an injectable prop
+ *
+ * @param responsiveStyles - The responsive styles.
+ *
+ * @returns An object expression representing the responsive styles.
+ */
+export const buildInjectableProp = (responsiveStyles: ResponsiveStyles) => {
+  const responsiveProperties = Object.entries(
+    responsiveStyles
+  ).map(([key, value]) => buildObjectProperty(key, t.arrayExpression(value)))
+
+  return t.objectExpression(responsiveProperties)
+}
